% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{rglimpse2_is_error}
\alias{rglimpse2_is_error}
\title{Test whether a value is an RGlimpse2 operational error}
\usage{
rglimpse2_is_error(value)
}
\arguments{
\item{value}{Any R value.}
}
\value{
A logical scalar.
}
\description{
Test whether a value is an RGlimpse2 operational error
}
