% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executables.R
\name{RGlimpse2Executables}
\alias{RGlimpse2Executables}
\title{Explicit GLIMPSE2 child executables}
\usage{
RGlimpse2Executables(
  chunk = character(0),
  split_reference = character(0),
  phase = character(0),
  ligate = character(0),
  phase_backend = character(0)
)
}
\arguments{
\item{chunk}{Absolute path to \code{GLIMPSE2_chunk}.}

\item{split_reference}{Absolute path to \code{GLIMPSE2_split_reference}.}

\item{phase}{Absolute path to the selected \code{GLIMPSE2_phase} executable.}

\item{ligate}{Absolute path to \code{GLIMPSE2_ligate}.}

\item{phase_backend}{Backend represented by \code{phase}.}
}
\description{
Explicit GLIMPSE2 child executables
}
